+++ Pandora/Paradigm/Structure.hs
@@ -14,7 +14,7 @@ import Pandora.Pattern.Functor.Pointable (point)
 import Pandora.Pattern.Transformer.Liftable (lift)
 import Pandora.Pattern.Object.Semigroup ((+))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, (||=))
-import Pandora.Paradigm.Inventory.Optics ((|>))
+import Pandora.Paradigm.Inventory.Optics ()
 import Pandora.Paradigm.Inventory.Store (Store (Store))
 import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True, False))
 import Pandora.Paradigm.Primary.Functor.Identity (Identity (Identity))
@@ -86,4 +86,4 @@ instance Accessible a (s :*: a) where
 	access = PQ_ $ \(s :*: x) -> Store $ x :*: (s :*:)
 
 instance {-# OVERLAPS #-} Accessible b a => Accessible b (s :*: a) where
-	access = access @a |> access @b
+	access = access @b . access @a