+++ Pandora/Paradigm/Structure/Ability/Substructure.hs
@@ -10,7 +10,7 @@ import Pandora.Pattern.Functor.Pointable (point)
 import Pandora.Pattern.Functor.Divariant ((>->))
 import Pandora.Pattern.Transformer.Liftable (lift)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, unite)
-import Pandora.Paradigm.Inventory.Optics (type (:~.), (|>))
+import Pandora.Paradigm.Inventory.Optics (type (:~.))
 import Pandora.Paradigm.Primary.Functor.Tagged (Tagged)
 import Pandora.Paradigm.Schemes.TU (type (<:.>))
 import Pandora.Paradigm.Schemes.PQ_ (PQ_ (PQ_))
@@ -28,6 +28,6 @@ data (|>) (i :: * -> k) (j :: * -> k') a
 
 instance (Covariant t, Covariant (Substructural i t), Substructure i t, Substructure j (Substructural i t)) => Substructure (i |> j) t where
 	type Substructural (i |> j) t = Substructural j (Substructural i t)
-	substructure = PQ_ $ extract . run >-> (unite . point <$>) $ run # sub @i |> sub @j
+	substructure = PQ_ $ extract . run >-> (unite . point <$>) $ run # sub @j . sub @i
 
 type Substructured i source target = (Substructure i source, Substructural i source ~ target)