+++ CHANGELOG.md
@@ -422,3 +422,4 @@
 * Add infix version of `hoist` method in `Hoistable` typeclass - /|\
 * Define `PQ_` joint schema exclusively for `Lens` type
 * Change `Lens` representation - wrap it in `PQ_` joint schema
+* Remove `|>` type operator in favor of `.` `Category` method